Deduplicate test setup and move factory to testing.py
- Move create_agent_with_events_dir from conftest.py to cli/testing.py
  (factory functions belong in testing.py, not conftest.py per CLAUDE.md)
- Add create_agent_with_sample_transcript helper that creates an agent
  with a pre-populated common_transcript events file
- Deduplicate integration tests in transcript_test.py by using the
  shared helpers instead of repeating event data inline
